Fix recurrence and prevent event edits

+++ app/controllers/discourse_events_integration/event_controller.rb
@@ -25,25 +25,32 @@ module DiscourseEventsIntegration
 
     def destroy
       event_ids = params[:event_ids]
-      destroy_topics = ActiveRecord::Type::Boolean.new.cast(params[:destroy_topics])
-      result = []
+      target = params[:target]
+      result = { destroyed_event_ids: [], destroyed_topics_event_ids: [] }
 
       ActiveRecord::Base.transaction do
         events = Event.where(id: event_ids)
 
-        if destroy_topics
+        if target === 'events_and_topics' || target === 'topics_only'
           events.includes(:event_connections).each do |event|
             event.event_connections.each do |ec|
-              PostDestroyer.new(current_user, ec.post).destroy
+              topic_id = ec.post.topic.id
+              destroyer = PostDestroyer.new(current_user, ec.post)
+              destroyer.destroy
             end
+
+            result[:destroyed_topics_event_ids] << event.id
           end
         end
 
-        result = events.destroy_all
+        if target === 'events_only' || target === 'events_and_topics'
+          destroyed_events = events.destroy_all
+          result[:destroyed_event_ids] += destroyed_events.map(&:id)
+        end
       end
 
-      if result.present?
-        render json: success_json.merge(destroyed_ids: result.map(&:id))
+      if result[:destroyed_event_ids].present? || result[:destroyed_topics_event_ids].present?
+        render json: success_json.merge(result)
       else
         render json: failed_json
       end
